ktlint

+++ core/src/main/kotlin/com/github/spigotbasics/core/storage/backends/MySQLBackend.kt
@@ -6,19 +6,27 @@ import com.github.spigotbasics.core.storage.StorageType
 import com.google.gson.JsonElement
 import java.util.concurrent.CompletableFuture
 
-internal class MySQLBackend(dbInfo: MySQLDatabaseInfo, sqlSleep: Double): HikariBackend(HikariConfigFactory.createMysqlConfig(dbInfo), dbInfo.tablePrefix, sqlSleep) {
-
+internal class MySQLBackend(dbInfo: MySQLDatabaseInfo, sqlSleep: Double) : HikariBackend(
+    HikariConfigFactory.createMysqlConfig(dbInfo),
+    dbInfo.tablePrefix,
+    sqlSleep,
+) {
     override val type = StorageType.MYSQL
 
-    override fun setJsonElement(namespace: String, keyId: String, value: JsonElement?): CompletableFuture<Void?> {
+    override fun setJsonElement(
+        namespace: String,
+        keyId: String,
+        value: JsonElement?,
+    ): CompletableFuture<Void?> {
         return CompletableFuture.runAsync {
             try {
                 selectSleep()
-                val sql = if (value == null) {
-                    "DELETE FROM $tablePrefix$namespace WHERE key_id = ?"
-                } else {
-                    "INSERT INTO $tablePrefix$namespace (key_id, data) VALUES (?, ?) ON DUPLICATE KEY UPDATE data = VALUES(data)"
-                }
+                val sql =
+                    if (value == null) {
+                        "DELETE FROM $tablePrefix$namespace WHERE key_id = ?"
+                    } else {
+                        "INSERT INTO $tablePrefix$namespace (key_id, data) VALUES (?, ?) ON DUPLICATE KEY UPDATE data = VALUES(data)"
+                    }
 
                 dataSource.connection.use { conn ->
                     conn.prepareStatement(sql).use { stmt ->
@@ -34,5 +42,4 @@ internal class MySQLBackend(dbInfo: MySQLDatabaseInfo, sqlSleep: Double): Hikari
             }
         }
     }
-
 }
